requirements.txt
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2023.11.17
charset-normalizer==3.3.2
h11==0.14.0
idna==3.6
lxml==5.3.0
outcome==1.3.0.post0
PySocks==1.7.1
requests==2.31.0
selenium==4.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.6
trio==0.24.0
trio-websocket==0.11.1
undetected-chromedriver==3.5.4
urllib3==2.1.0
websockets==12.0
wsproto==1.2.0
setuptools
google.generativeai
boto3==1.34.34
python-dotenv==1.0.1
nicegui==1.4.6
streamlit
plotly
deepdiff
tomli